specify sizes of uints for cache size variables
addresses review comments https://github.com/ava-labs/awm-relayer/pull/415#discussion_r1711851055
and https://github.com/ava-labs/awm-relayer/pull/415#discussion_r1711855293

+++ signature-aggregator/aggregator/aggregator.go
@@ -65,7 +65,7 @@ type SignatureAggregator struct {
 func NewSignatureAggregator(
 	network *peers.AppRequestNetwork,
 	logger logging.Logger,
-	signatureCacheSize uint,
+	signatureCacheSize uint64,
 	metrics *metrics.SignatureAggregatorMetrics,
 	messageCreator message.Creator,
 ) (*SignatureAggregator, error) {
